+++ src/cli/task-serve.ts
@@ -6,14 +6,18 @@ import exit from 'exit';
 export async function taskServe(process: NodeJS.Process, config: d.Config, flags: d.ConfigFlags) {
   const { Compiler } = require('../compiler/index.js');
 
+  config.suppressLogs = true;
+
   const compiler: d.Compiler = new Compiler(config);
   if (!compiler.isValid) {
     exit(1);
   }
 
   config.flags.serve = true;
-  config.devServer.openBrowser = false;
+  config.devServer.openBrowser = flags.open;
   config.devServer.hotReplacement = false;
+  config.devServer.initialLoadUrl = '/';
+  config.devServer.websocket = false;
   config.maxConcurrentWorkers = 1;
 
   config.devServer.root = process.cwd();