+++ src/compiler/copy/copy-tasks-worker.ts
@@ -18,9 +18,13 @@ export async function copyTasksWorker(copyTasks: d.CopyTask[]) {
 
     // figure out all the file copy tasks we'll have
     // by digging down through any directory copy tasks
-    await Promise.all(copyTasks.map(async copyTask => {
-      await processCopyTask(fs, results, allCopyTasks, copyTask);
-    }));
+    while (copyTasks.length > 0) {
+      const tasks = copyTasks.splice(0, 100);
+
+      await Promise.all(tasks.map(async copyTask => {
+        await processCopyTask(fs, results, allCopyTasks, copyTask);
+      }));
+    }
 
     // figure out which directories we'll need to make first
     const mkDirs = ensureDirs(allCopyTasks);
@@ -30,10 +34,13 @@ export async function copyTasksWorker(copyTasks: d.CopyTask[]) {
       } catch (mkDirErr) {}
     }
 
-    // begin copying all the files
-    await Promise.all(allCopyTasks.map(async copyTask => {
-      await fs.copyFile(copyTask.src, copyTask.dest);
-    }));
+    while (allCopyTasks.length > 0) {
+      const tasks = allCopyTasks.splice(0, 100);
+
+      await Promise.all(tasks.map(async copyTask => {
+        await fs.copyFile(copyTask.src, copyTask.dest);
+      }));
+    }
 
   } catch (e) {
     catchError(results.diagnostics, e);