+++ src/dev-server/serve-file.ts
@@ -15,7 +15,7 @@ export async function serveFile(devServerConfig: d.DevServerConfig, fs: d.FileSy
       // easy text file, use the internal cache
       let content = await fs.readFile(req.filePath);
 
-      if (util.isHtmlFile(req.filePath) && !util.isDevServerClient(req.pathname)) {
+      if (devServerConfig.websocket && util.isHtmlFile(req.filePath) && !util.isDevServerClient(req.pathname)) {
         // auto inject our dev server script
         content += getDevServerClientScript(devServerConfig, req);