+++ src/mock-doc/location.ts
@@ -30,6 +30,7 @@ export class MockLocation implements Location {
     this.username = url.username;
     this.password = url.password;
     this.origin = url.origin;
+    this._href = value;
   }
 
   assign(_url: string) {